Roadmap updates per meeting this week
This is an attempt to update and re-organize the roadmap. I am sure things are not where the maintainers would like them, so please comment and I will adjust accordingly.

## Which problem is this PR solving?
Roadmap Updates

## Short description of the changes
Re-organized existing roadmap, and added several items which are in active development today. Added more to the wish list and longer term goals.

Signed-off-by: Jonah Kowall <jkowall@kowall.net>