fix scrolling issue on long page tocs and left nav
Fixes #360 where long tocs were defaulting to overflow: hidden in the
theme. This commit adds a CSS file to override the default, which
sets the overflow to scroll.

Fixes #508. Adds similar styles to left nav styles so that they are
scrollable. Also adds 1rem to max-width because with the scrollbar,
the word "Troubleshooting" gets cut off.

Signed-off-by: Kapunahele Wong <kapunahelewong@gmail.com>