Remove shell completion configs
We probably upgraded to a version of viper/cobra libraries that
started generating shell completion configs alongside with the
CLI flags dumps. We're not using these files for the website,
so this change deletes those that were already checked in, and
fixes the release script to remove them automatically in the future.

Signed-off-by: Yuri Shkuro <github@ysh.us>