Fix page cross-version links, drop canonifyURLs, and more (#891)
- Contributes to #889
- Resolves invalid links of the form
`https://www.jaegertracing.io/docs/2.6/(cli|operator)/`
- Fixes `themes/jaeger-docs/layouts/partials/docs/header.html` so that
it doesn't create invalid cross-version links
- Drops the now deprecated
[canonifyURLs](https://gohugo.io/content-management/urls/#canonical-urls)
Hugo config option, since it unnecessarily creates external links to
intra-site targets, that should be encoded as simple paths
- Fixes other misc links and hash targets
- Adjusts some layouts so that local paths are used for intra-site
links. When a full URL is required, we've added `data-proofer-ignore` so
that the link checker doesn't need to be bothered with it.

---------

Signed-off-by: Patrice Chalin <pchalin@gmail.com>