Navbar partial whitespace cleanup (#901)
- Contributes to #746
- There are **no change to the generated production (i.e., minified)
site files**
- Trims superfluous whitespace from
`themes/jaeger-docs/layouts/partials/navbar.html`. This partial iterates
over all docs, and for each iteration adds whitespace to the navbar.
This creates files with _a lot_ of unnecessary blank lines, which makes
it all the more challenging to diff the site-file changes necessary in
preparation for the Docsy migration.
- In most cases, a Hugo template code statement ending in `}}` is
replaced by `-}}`; a directive that will trim the excess whitespace
following the directive when the page is generated.

Signed-off-by: Patrice Chalin <pchalin@gmail.com>